nmz: refactor(PrayerPotionScript): move script to nmz package and simplify method signature
